requirements.txt
networkx
asn1crypto==0.24.0
blinker==1.4
boto3==1.9.134
botocore==1.12.134
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
docutils==0.14
future==0.17.1
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
python-twitter==3.5
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.3
win-inet-pton==1.1.0
wincertstore==0.2
